refactor: move routing & superuser to base controller
and make every custom controller extends base controller
also require every controller to extend base the controller, no more
"maybe it will work", no more warnings
