+++ app/controllers/v1/files.ts
@@ -7,6 +7,7 @@ import type { Constructor, LazyImport } from "@adonisjs/core/types/http";
 import drive from "@adonisjs/drive/services/main";
 import { errors as limiterErrors } from "@adonisjs/limiter";
 
+import BaseController from "#controllers/base_controller";
 import {
   BadRequestException,
   TooManyRequestsException,
@@ -34,7 +35,7 @@ const directUploadValidator = vine.compile(
   }),
 );
 
-export default class FilesController {
+export default class FilesController extends BaseController {
   $configureRoutes(controller: LazyImport<Constructor<FilesController>>) {
     router.get("/:key", [controller, "get"]).as("show");
     router.post("/", [controller, "post"]).as("upload").use(middleware.auth());
